添加设置精华

+++ src/Model/Content/DeleteMsgContent.cs
@@ -3,9 +3,9 @@ using Milimoe.OneBot.Framework.Interface;
 
 namespace Milimoe.OneBot.Model.Content
 {
-    public class DeleteMsgContent(int message_id) : IContent
+    public class DeleteMsgContent(long message_id) : IContent
     {
-        public int message_id { get; set; } = message_id;
+        public long message_id { get; set; } = message_id;
 
         [JsonIgnore]
         public string detail => "撤回消息：" + message_id;
