+++ src/Model/Content/EssenceMsgContent.cs
@@ -3,9 +3,9 @@ using Milimoe.OneBot.Framework.Interface;
 
 namespace Milimoe.OneBot.Model.Content
 {
-    public class EssenceMsgContent(int message_id) : IContent
+    public class EssenceMsgContent(long message_id) : IContent
     {
-        public int message_id { get; set; } = message_id;
+        public long message_id { get; set; } = message_id;
 
         [JsonIgnore]
         public string detail => "设置精华：" + message_id;
